# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/customer/{slug}/panel/funding/initiate:
    post:
      operationId: initiateCustomerPayment
      tags: [Payments]
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/FundingCreateInput"}
      responses:
        "202":
          description: Durable initiation accepted; not proof of payment or wallet credit
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FundingView"}
  /payments/webhooks/{code}/{binding}:
    post:
      operationId: receiveProviderWebhook
      tags: [Payments]
      description: Raw bytes authenticated exclusively by a trusted installed provider verifier. No browser/session confirmation.
      parameters:
        - {name: code, in: path, required: true, schema: {type: string}}
        - {name: binding, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: {type: string, format: binary}
      responses:
        "202":
          description: Authenticated evidence durably queued; does not imply settlement
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WebhookAcknowledgement"}
...
  /public/customer/{slug}/panel/funding:
    get:
      operationId: listCustomerFundingRequests
      tags: [Payments]
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Own funding history, separate from account statement
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FundingList"}
    post:
      operationId: createCustomerFundingRequest
      tags: [Payments]
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/FundingCreateInput"}
      responses:
        "201":
          description: Frozen funding intent; not payment confirmation
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FundingView"}
  /public/customer/{slug}/panel/funding/gateways:
    get:
      operationId: listCustomerFundingGateways
      tags: [Payments]
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Only operational, globally and tenant enabled, compatible gateways
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GatewayList"}
  /public/customer/{slug}/panel/funding/quote:
    post:
      operationId: quoteCustomerFunding
      tags: [Payments]
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/FundingIntent"}
      responses:
        "200":
          description: Exact server-calculated preview; creation freezes final snapshot
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FundingQuote"}
  /public/customer/{slug}/panel/funding/{id}:
    get:
      operationId: getCustomerFundingRequest
      tags: [Payments]
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Own funding request and safe payment status
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FundingView"}
  /public/customer/{slug}/panel/funding/{id}/cancel:
    post:
      operationId: cancelCustomerFundingRequest
      tags: [Payments]
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/EmptyPaymentAction"}
      responses:
        "200":
          description: Unpaid funding request cancelled; no wallet mutation
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FundingView"}
...
  /public/customer/{slug}/panel:
    parameters: [{in: path, name: slug, required: true, schema: {type: string}}]
    get:
      operationId: getCustomerPanel
      responses:
        '200': {description: Own financial summary and real service-order metrics, content: {application/json: {schema: {$ref: '#/components/schemas/FoundationResponse'}}}}
  /public/customer/{slug}/panel/services:
    parameters: [{in: path, name: slug, required: true, schema: {type: string}}]
    get:
      operationId: listCustomerServices
      parameters:
        - {$ref: '#/components/parameters/FoundationPage'}
        - {$ref: '#/components/parameters/FoundationSearch'}
        - {$ref: '#/components/parameters/FoundationServiceType'}
        - {in: query, name: groupId, schema: {type: string, format: uuid}}
        - {in: query, name: currency, schema: {type: string, minLength: 3, maxLength: 3}}
      responses:
        '200': {description: Active tenant services with exact formatted prices, content: {application/json: {schema: {$ref: '#/components/schemas/FoundationResponse'}}}}
  /public/customer/{slug}/panel/services/{id}:
    parameters: [{in: path, name: slug, required: true, schema: {type: string}}, {in: path, name: id, required: true, schema: {type: string, format: uuid}}]
    get:
      operationId: getCustomerService
      responses:
        '200': {description: Active service and requirement fields, content: {application/json: {schema: {$ref: '#/components/schemas/FoundationResponse'}}}}
  /public/customer/{slug}/panel/quote:
    parameters: [{in: path, name: slug, required: true, schema: {type: string}}]
    post:
      operationId: quoteCustomerService
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ServiceQuoteInput'}}}}
      responses:
        '200': {description: Server price and sufficient-funds assessment, content: {application/json: {schema: {$ref: '#/components/schemas/FoundationResponse'}}}}
  /public/customer/{slug}/panel/orders:
    parameters: [{in: path, name: slug, required: true, schema: {type: string}}]
    get:
      operationId: listCustomerServiceOrders
      parameters:
        - {$ref: '#/components/parameters/FoundationPage'}
        - {$ref: '#/components/parameters/FoundationSearch'}
        - {$ref: '#/components/parameters/FoundationServiceType'}
        - {in: query, name: status, schema: {type: string, enum: [pending, processing, completed, rejected, cancelled]}}
      responses:
        '200': {description: Own service orders, content: {application/json: {schema: {$ref: '#/components/schemas/FoundationResponse'}}}}
    post:
      operationId: placeCustomerServiceOrder
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ServicePurchaseInput'}}}}
      responses:
        '201': {description: Atomic idempotent order and debit, content: {application/json: {schema: {$ref: '#/components/schemas/FoundationResponse'}}}}
        '409': {description: Insufficient balance or conflicting idempotency key}
  /public/customer/{slug}/panel/orders/{id}:
    parameters: [{in: path, name: slug, required: true, schema: {type: string}}, {in: path, name: id, required: true, schema: {type: string, format: uuid}}]
    get:
      operationId: getCustomerServiceOrder
      responses:
        '200': {description: Own order excluding reseller notes, content: {application/json: {schema: {$ref: '#/components/schemas/FoundationResponse'}}}}
  /public/customer/{slug}/panel/statement:
    parameters: [{in: path, name: slug, required: true, schema: {type: string}}]
    get:
      operationId: getCustomerStatement
      parameters:
        - {$ref: '#/components/parameters/FoundationPage'}
        - {$ref: '#/components/parameters/FoundationSearch'}
        - {in: query, name: direction, schema: {type: string, enum: [credit, debit]}}
        - {in: query, name: type, schema: {type: string, enum: [admin_credit, admin_debit, adjustment, order_debit, order_refund, payment_credit]}}
      responses:
        '200': {description: Own paginated statement excluding internal notes and actor IDs, content: {application/json: {schema: {$ref: '#/components/schemas/FoundationResponse'}}}}
...
  /public/customer/{slug}/panel/security:
    get:
      operationId: getCustomerSecurity
      tags: [PublicCustomer]
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Own security metadata
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CustomerSecurityView"}
  /public/customer/{slug}/panel/security/password:
    post:
      operationId: changeCustomerPassword
      tags: [PublicCustomer]
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CustomerPasswordChangeInput"}
      responses:
        "200":
          description: Password changed; current session retained
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CustomerSecurityResult"}
  /public/customer/{slug}/panel/security/sessions/{session}/revoke:
    post:
      operationId: revokeCustomerSession
      tags: [PublicCustomer]
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
        - {name: session, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CustomerSecurityAction"}
      responses:
        "200":
          description: Owned non-current session revoked
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CustomerSecurityResult"}
  /public/customer/{slug}/panel/security/sessions/revoke-others:
    post:
      operationId: revokeOtherCustomerSessions
      tags: [PublicCustomer]
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CustomerSecurityAction"}
      responses:
        "200":
          description: Other sessions revoked
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CustomerSecurityResult"}
...
  /subscribers/{id}/panel:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: getSubscriberPanel
      responses:
        "200":
          description: Tenant checked and licence allowed; business modules remain deferred
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PanelAccess" }
        "403":
          description: Tenant mismatch or pending/suspended/expired/revoked licence
...
components:
  parameters:
    FoundationPage: {in: query, name: page, schema: {type: integer, minimum: 1, maximum: 100000, default: 1}}
    FoundationSearch: {in: query, name: search, schema: {type: string, maxLength: 100}}
    FoundationServiceType: {in: query, name: serviceType, schema: {type: string, enum: [imei, server, file, remote]}}
  schemas:
    EmptyPaymentAction:
      type: object
      additionalProperties: false
...
    FundingCreateInput:
      allOf:
        - {$ref: "#/components/schemas/FundingIntent"}
        - {type: object, required: [idempotencyKey], properties: {idempotencyKey: {type: string, format: uuid}}}
    GatewayList:
      type: object
      required: [data, credentialStorageReady]
      properties:
        data: {type: array, items: {$ref: "#/components/schemas/GatewayView"}}
        credentialStorageReady: {type: boolean}
    GatewayField:
      type: object
      required: [key, label, secret, required]
      properties:
        key: {type: string}
        label: {type: string}
        secret: {type: boolean}
        required: {type: boolean}
    GatewayCurrencyRule:
      type: object
      required: [currency, minimum, maximum, feeBps, fixedFee]
      properties:
        currency: {type: string}
        minimum: {type: string}
        maximum: {type: string}
        feeBps: {type: integer, minimum: 0, maximum: 10000}
        fixedFee: {type: string}
...
    GatewayView:
      type: object
      required: [code, displayName, description, supportedCurrencies, supportedMethods, automaticConfirmationSupported, webhookSupported, requiredCredentials, configurationFields, integrationStatus, version, feesSupported, globalEnabled, resellerAvailable, operational, enabled, instructions, currencyRules, configuredCredentialFields, validationStatus]
      properties:
        capabilities: {$ref: "#/components/schemas/PaymentCapabilities"}
        callbackPath: {type: [string, "null"]}
        code: {type: string}
        displayName: {type: string}
        description: {type: string}
        supportedCurrencies: {type: array, items: {type: string}}
        configurationSchema: {type: object, additionalProperties: true}
        supportedMethods: {type: array, items: {type: string}}
        automaticConfirmationSupported: {type: boolean}
        webhookSupported: {type: boolean}
        requiredCredentials: {type: array, items: {$ref: "#/components/schemas/GatewayField"}}
        configurationFields: {type: array, items: {$ref: "#/components/schemas/GatewayField"}}
        integrationStatus: {type: string, enum: [AVAILABLE, NOT_IMPLEMENTED, DISABLED]}
        version: {type: string}
        feesSupported: {type: boolean}
        globalEnabled: {type: boolean}
        resellerAvailable: {type: boolean}
        operational: {type: boolean}
        enabled: {type: boolean}
        instructions: {type: string}
        currencyRules: {type: array, items: {$ref: "#/components/schemas/GatewayCurrencyRule"}}
        configuredCredentialFields: {type: array, items: {type: string}}
        validationStatus: {type: string, enum: [NOT_IMPLEMENTED, NOT_CONFIGURED, NOT_VALIDATED, VALID, INVALID]}
    FundingIntent:
      type: object
      required: [gatewayCode, paymentMethod, amount, paymentCurrency]
      properties:
        gatewayCode: {type: string}
        paymentMethod: {type: string}
        amount: {type: string, description: Requested wallet credit in fixed account currency}
        paymentCurrency: {type: string}
    FundingQuote:
      type: object
      required: [accountCurrency, paymentCurrency, requestedCreditUnits, paymentBaseMinor, feeMinor, expectedPaymentMinor, formattedCredit, formattedBase, formattedFee, formattedPayable, fxDescription]
      properties:
        accountCurrency: {type: string}
        paymentCurrency: {type: string}
        requestedCreditUnits: {type: string}
        paymentBaseMinor: {type: string}
        feeMinor: {type: string}
        expectedPaymentMinor: {type: string}
        formattedCredit: {type: string}
        formattedBase: {type: string}
        formattedFee: {type: string}
        formattedPayable: {type: string}
        fxDescription: {type: string}
    FundingPayment:
      type: object
      required: [id, status, amountMinor, formattedAmount, currency, providerReference, createdAt, settledAt]
      properties:
        id: {type: string, format: uuid}
        status: {type: string, enum: [VERIFIED, SETTLED, FAILED]}
        amountMinor: {type: string}
        formattedAmount: {type: string}
        currency: {type: string}
        providerReference: {type: [string, "null"]}
        createdAt: {type: string, format: date-time}
        settledAt: {type: [string, "null"], format: date-time}
    FundingView:
      allOf:
        - {$ref: "#/components/schemas/FundingQuote"}
        - type: object
          required: [id, gatewayCode, gatewayName, paymentMethod, customerName, status, createdAt, updatedAt, expiresAt, payments, paidAt, creditedWalletAmount, reviewRequired, reviewReasons, settlementStatus, processing]
          properties:
            paidAt: {type: [string, "null"], format: date-time}
            creditedWalletAmount: {type: [string, "null"]}
            reviewRequired: {type: boolean}
            reviewReasons: {type: array, items: {type: string}}
            settlementStatus: {type: string, enum: [NONE, VERIFIED, SETTLED, REVIEW_REQUIRED, FAILED]}
            processing: {$ref: "#/components/schemas/FundingProcessing"}
            id: {type: string, format: uuid}
            gatewayCode: {type: string}
            gatewayName: {type: string}
            paymentMethod: {type: string}
            customerName: {type: [string, "null"]}
            status: {type: string, enum: [CREATED, PENDING_PAYMENT, PAID, FAILED, EXPIRED, CANCELLED]}
            createdAt: {type: string, format: date-time}
            updatedAt: {type: string, format: date-time}
            expiresAt: {type: string, format: date-time}
            payments: {type: array, items: {$ref: "#/components/schemas/FundingPayment"}}
    FundingList:
      type: object
      required: [data, hasMore]
      properties:
        data: {type: array, items: {$ref: "#/components/schemas/FundingView"}}
        hasMore: {type: boolean}
    PaymentCapabilities:
      type: object
      required: [createPayment, idempotentCreation, statusLookup, webhookVerification, cancelPayment, refundPayment]
      properties:
        createPayment: {type: boolean}
        idempotentCreation: {type: boolean}
        statusLookup: {type: boolean}
        webhookVerification: {type: boolean}
        cancelPayment: {type: boolean}
        refundPayment: {type: boolean}
...
    FundingProcessing:
      type: [object, "null"]
      required: [state, paymentUrl, paymentAddress, instructions, expiresAt]
      properties:
        state: {type: string, enum: [READY, PROCESSING, REVIEW_REQUIRED, FAILED]}
        paymentUrl: {type: [string, "null"]}
        paymentAddress: {type: [string, "null"]}
        instructions: {type: string}
        expiresAt: {type: [string, "null"], format: date-time}
...
    WebhookAcknowledgement:
      type: object
      required: [received, authenticated, queued, duplicate]
      properties:
        received: {type: boolean}
        authenticated: {type: boolean}
        queued: {type: boolean}
        duplicate: {type: boolean}
    CustomerSecurityAction:
      type: object
      additionalProperties: false
    CustomerSecurityResult:
      type: object
      required: [ok]
      properties:
        ok: {type: boolean}
        affected: {type: integer}
    CustomerPasswordChangeInput:
      type: object
      required: [currentPassword, newPassword, confirmPassword]
      additionalProperties: false
      properties:
        currentPassword: {type: string, minLength: 1, maxLength: 128}
        newPassword: {type: string, minLength: 8, maxLength: 128}
        confirmPassword: {type: string, minLength: 8, maxLength: 128}
...
    CustomerSecurityView:
      type: object
      required: [enabled, lastLoginAt, lastLoginIp, lastLoginSource, passwordChangedAt, lockedUntil, activeSessionCount, sessions, history]
      properties:
        enabled: {type: boolean}
        lastLoginAt: {type: [string, "null"], format: date-time}
        lastLoginIp: {type: [string, "null"]}
        lastLoginSource: {type: [string, "null"], enum: [history, legacy, null]}
        passwordChangedAt: {type: [string, "null"], format: date-time}
        lockedUntil: {type: [string, "null"], format: date-time}
        activeSessionCount: {type: integer}
        sessions:
          type: array
          items: {$ref: "#/components/schemas/CustomerSecuritySession"}
        history:
          type: array
          items: {$ref: "#/components/schemas/CustomerLoginAttempt"}
    CustomerSecuritySession:
      type: object
      required: [id, current, device, ipAddress, createdAt, lastSeenAt, expiresAt, status]
      properties:
        id: {type: string, format: uuid}
        current: {type: boolean}
        device: {type: string}
        ipAddress: {type: [string, "null"]}
        createdAt: {type: string, format: date-time}
        lastSeenAt: {type: [string, "null"], format: date-time}
        expiresAt: {type: string, format: date-time}
        status: {type: string, enum: [active]}
    CustomerLoginAttempt:
      type: object
      required: [id, createdAt, result, ipAddress, device]
      properties:
        id: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
        result: {type: string, enum: [success, failed, locked, blocked]}
        ipAddress: {type: [string, "null"]}
        device: {type: string}
    FoundationResponse:
      type: object
      description: Foundation JSON envelope; exact USD units are decimal strings, never floating-point amounts.
      additionalProperties: {}
...
    ServiceQuoteInput:
      type: object
      required: [serviceId]
      additionalProperties: false
      properties:
        serviceId: {type: string, format: uuid}
        currency: {type: string, minLength: 3, maxLength: 3}
    ServicePurchaseInput:
      type: object
      additionalProperties: false
      required: [serviceId, expectedPriceUsdUnits, idempotencyKey, inputs]
      properties:
        serviceId: {type: string, format: uuid}
        currency: {type: string, minLength: 3, maxLength: 3}
        expectedPriceUsdUnits: {type: string, pattern: '^[0-9]{1,24}$', description: Source-price stale-quote guard. Actual charge is server-calculated in the fixed client account currency.}
        expectedPriceAccountUnits: {type: string, pattern: '^[0-9]{1,24}$', description: Account-currency stale-quote guard; required for non-USD clients.}
        idempotencyKey: {type: string, format: uuid}
        inputs: {type: object, additionalProperties: {type: string, maxLength: 4000}}
...
    PanelAccess:
      type: object
      required: [subscriberId, allowed]
      properties:
        subscriberId: { type: string }
        allowed: { type: boolean }
...