# lib/api-zod/src/generated/types/gatewayView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GatewayCurrencyRule } from './gatewayCurrencyRule';
import type { GatewayField } from './gatewayField';
import type { GatewayViewConfigurationSchema } from './gatewayViewConfigurationSchema';
import type { GatewayViewIntegrationStatus } from './gatewayViewIntegrationStatus';
import type { GatewayViewValidationStatus } from './gatewayViewValidationStatus';
import type { PaymentCapabilities } from './paymentCapabilities';

export interface GatewayView {
  capabilities?: PaymentCapabilities;
  /** @nullable */
  callbackPath?: string | null;
  code: string;
  displayName: string;
  description: string;
  supportedCurrencies: string[];
  configurationSchema?: GatewayViewConfigurationSchema;
  supportedMethods: string[];
  automaticConfirmationSupported: boolean;
  webhookSupported: boolean;
  requiredCredentials: GatewayField[];
  configurationFields: GatewayField[];
  integrationStatus: GatewayViewIntegrationStatus;
  version: string;
  feesSupported: boolean;
  globalEnabled: boolean;
  resellerAvailable: boolean;
  operational: boolean;
  enabled: boolean;
  instructions: string;
  currencyRules: GatewayCurrencyRule[];
  configuredCredentialFields: string[];
  validationStatus: GatewayViewValidationStatus;
}